+++ packages/xchain-cosmos/src/client.ts
@@ -240,8 +240,6 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
       asset: getDenom(asset || mainAsset),
       memo,
     })
-
-    // return transferResult || ''
   }
 
   /**