+++ packages/xchain-cosmos/src/cosmos/sdk-client.ts
@@ -249,8 +249,6 @@ export class CosmosSDKClient {
       fee,
     })
 
-    console.log('txBody: ', txBody)
-
     const txBuilder = new cosmosclient.TxBuilder(this.sdk, txBody, authInfo)
 
     return this.signAndBroadcast(txBuilder, privkey, account)
@@ -303,21 +301,17 @@ export class CosmosSDKClient {
     const signDocBytes = txBuilder.signDocBytes(signerAccount.account_number)
     txBuilder.addSignature(privKey.sign(signDocBytes))
 
-    console.log('tx payload: ', {
-      tx_bytes: txBuilder.txBytes(),
-      mode: rest.tx.BroadcastTxMode.Block,
-    })
-
     // broadcast
     const res = await rest.tx.broadcastTx(this.sdk, {
       tx_bytes: txBuilder.txBytes(),
-      mode: rest.tx.BroadcastTxMode.Block,
+      mode: rest.tx.BroadcastTxMode.Sync,
     })
 
     if (res?.data?.tx_response?.code !== 0) {
       throw new Error('Error broadcasting: ' + res?.data?.tx_response?.raw_log)
     }
 
-    return res.data.tx_response.txhash || ''
+    if (!res.data.tx_response.txhash) throw new Error('TxHash is missing or invalid')
+    return res.data.tx_response.txhash
   }
 }