+++ packages/xchain-thorchain/__tests__/client.test.ts
@@ -1,429 +1,514 @@
-// import { Network, TxsPage } from '@thorswap-lib/xchain-client'
-// import { CosmosSDKClient, RPCResponse, RPCTxSearchResult, TxResponse } from '@thorswap-lib/xchain-cosmos'
-// import { AssetRuneNative, BaseAmount, assetAmount, assetToBase, baseAmount } from '@thorswap-lib/xchain-util'
-// import { BaseAccount, BroadcastTxCommitResult, Coin } from 'cosmos-client/api'
-// import nock from 'nock'
+import { Coin } from '@cosmos-client/core/esm/openapi/api'
+import { Network, TxsPage } from '@thorswap-lib/xchain-client'
+import { CosmosSDKClient, RPCResponse, RPCTxSearchResult, TxResponse } from '@thorswap-lib/xchain-cosmos'
+import {
+  // Asset,
+  // AssetBNB,
+  AssetRuneNative,
+  // BNBChain,
+  BaseAmount,
+  assetAmount,
+  assetToBase,
+  baseAmount,
+} from '@thorswap-lib/xchain-util'
+import nock from 'nock'
 
-// import { Client } from '../src/client'
+import { mockTendermintNodeInfo } from '../__mocks__/thornode-api'
+import { Client } from '../src/client'
 // import { ThorchainDepositResponse } from '../src/types'
 
-// const mockAccountsAddress = (
-//   url: string,
-//   address: string,
-//   result: {
-//     height: number
-//     result: BaseAccount
-//   },
-// ) => {
-//   nock(url).get(`/auth/accounts/${address}`).reply(200, result)
-// }
+const mockAccountsAddress = (
+  url: string,
+  address: string,
+  result: {
+    account: {
+      '@type': string
+      address: string
+      pub_key?: {
+        '@type': string
+        key: string
+      }
+      account_number: string
+      sequence: string
+    }
+  },
+) => {
+  nock(url).get(`/cosmos/auth/v1beta1/accounts/${address}`).reply(200, result)
+}
 
-// const mockAccountsBalance = (
-//   url: string,
-//   address: string,
-//   result: {
-//     height: number
-//     result: Coin[]
-//   },
-// ) => {
-//   nock(url).get(`/bank/balances/${address}`).reply(200, result)
-// }
+const mockAccountsBalance = (
+  url: string,
+  address: string,
+  result: {
+    balances: Coin[]
+  },
+) => {
+  nock(url).get(`/cosmos/bank/v1beta1/balances/${address}`).reply(200, result)
+}
+
+const mockThorchainConstants = (url: string) => {
+  const response = require('../__mocks__/responses/thorchain/constants.json')
+  nock(url).get('/thorchain/constants').reply(200, response)
+}
 
 // const mockThorchainDeposit = (url: string, result: ThorchainDepositResponse) => {
 //   nock(url).post('/thorchain/deposit').reply(200, result)
 // }
 
-// const assertTxsPost = (url: string, memo: undefined | string, result: BroadcastTxCommitResult): void => {
-//   nock(url)
-//     .post(`/txs`, (body) => {
-//       expect(body.tx.msg.length).toEqual(1)
-//       expect(body.tx.memo).toEqual(memo)
-//       return true
-//     })
-//     .reply(200, result)
-// }
+const assertTxsPost = (
+  url: string,
+  result: {
+    tx_response: {
+      txhash: string
+      code: number
+    }
+  },
+): void => {
+  nock(url, { allowUnmocked: true })
+    .post(`/cosmos/tx/v1beta1/txs`, (body) => {
+      expect(body.mode).toEqual('BROADCAST_MODE_SYNC')
+      expect(body.tx_bytes.length).toBeGreaterThan(0)
+      return true
+    })
+    .reply(200, result)
+}
 
-// const mockTxHistory = (url: string, result: RPCResponse<RPCTxSearchResult>): void => {
-//   nock(url)
-//     .get(`/tx_search`)
-//     .twice()
-//     .query((_) => true)
-//     .reply(200, result)
-// }
+const mockTxHistory = (url: string, result: RPCResponse<RPCTxSearchResult>): void => {
+  nock(url)
+    .get(`/tx_search`)
+    .twice()
+    .query((_) => true)
+    .reply(200, result)
+}
 
-// const assertTxHashGet = (url: string, hash: string, result: TxResponse): void => {
-//   nock(url).get(`/txs/${hash}`).reply(200, result)
-// }
+const assertTxHashGet = (url: string, hash: string, result: { tx_response: TxResponse }): void => {
+  nock(url).get(`/cosmos/tx/v1beta1/txs/${hash}`).reply(200, result)
+}
 
-// describe('Client Test', () => {
-//   let thorClient: Client
-//   let thorMainClient: Client
-//   const phrase = 'rural bright ball negative already grass good grant nation screen model pizza'
-//   const mainnet_address_path0 = 'thor19kacmmyuf2ysyvq3t9nrl9495l5cvktjs0yfws'
-//   const mainnet_address_path1 = 'thor1hrf34g3lxwvpk7gjte0xvahf3txnq8ecgaf4nc'
-//   const testnet_address_path0 = 'tthor19kacmmyuf2ysyvq3t9nrl9495l5cvktj5c4eh4'
-//   const testnet_address_path1 = 'tthor1hrf34g3lxwvpk7gjte0xvahf3txnq8ecv2c92a'
+describe('Client Test', () => {
+  let thorClient: Client
+  let thorMainClient: Client
+  const phrase = 'rural bright ball negative already grass good grant nation screen model pizza'
+  const mainnet_address_path0 = 'thor19kacmmyuf2ysyvq3t9nrl9495l5cvktjs0yfws'
+  const mainnet_address_path1 = 'thor1hrf34g3lxwvpk7gjte0xvahf3txnq8ecgaf4nc'
+  const testnet_address_path0 = 'tthor19kacmmyuf2ysyvq3t9nrl9495l5cvktj5c4eh4'
+  const testnet_address_path1 = 'tthor1hrf34g3lxwvpk7gjte0xvahf3txnq8ecv2c92a'
 
-//   beforeEach(() => {
-//     thorClient = new Client({ phrase, network: 'testnet' as Network })
-//     thorMainClient = new Client({ phrase, network: 'mainnet' as Network })
-//   })
+  beforeEach(() => {
+    thorClient = new Client({ phrase, network: Network.Testnet })
+    thorMainClient = new Client({ phrase, network: Network.Mainnet })
+  })
 
-//   afterEach(() => {
-//     thorClient.purgeClient()
-//     thorMainClient.purgeClient()
-//   })
+  afterEach(() => {
+    thorClient.purgeClient()
+    thorMainClient.purgeClient()
+  })
 
-//   it('should start with empty wallet', async () => {
-//     const thorClientEmptyMain = new Client({ phrase, network: 'mainnet' as Network })
-//     const addressMain = thorClientEmptyMain.getAddress()
-//     expect(addressMain).toEqual(mainnet_address_path0)
+  it('should start with empty wallet', async () => {
+    const thorClientEmptyMain = new Client({ phrase, network: 'mainnet' as Network })
+    const addressMain = thorClientEmptyMain.getAddress()
+    expect(addressMain).toEqual(mainnet_address_path0)
 
-//     const thorClientEmptyTest = new Client({ phrase, network: 'testnet' as Network })
-//     const addressTest = thorClientEmptyTest.getAddress()
-//     expect(addressTest).toEqual(testnet_address_path0)
-//   })
+    const thorClientEmptyTest = new Client({ phrase, network: 'testnet' as Network })
+    const addressTest = thorClientEmptyTest.getAddress()
+    expect(addressTest).toEqual(testnet_address_path0)
+  })
 
-//   it('should derive address accordingly to the user param', async () => {
-//     const thorClientEmptyMain = new Client({
-//       phrase,
-//       network: 'mainnet' as Network /*, derivationPath: "44'/931'/0'/0/0" */,
-//     })
-//     const addressMain = thorClientEmptyMain.getAddress()
-//     expect(addressMain).toEqual(mainnet_address_path0)
+  it('should derive address accordingly to the user param', async () => {
+    const thorClientEmptyMain = new Client({
+      phrase,
+      network: 'mainnet' as Network /*, derivationPath: "44'/931'/0'/0/0" */,
+    })
+    const addressMain = thorClientEmptyMain.getAddress()
+    expect(addressMain).toEqual(mainnet_address_path0)
 
-//     const viaSetPhraseAddr1 = thorClientEmptyMain.getAddress(1 /*, "44'/931'/0'/0/1" */)
-//     expect(viaSetPhraseAddr1).toEqual(mainnet_address_path1)
+    const viaSetPhraseAddr1 = thorClientEmptyMain.getAddress(1 /*, "44'/931'/0'/0/1" */)
+    expect(viaSetPhraseAddr1).toEqual(mainnet_address_path1)
 
-//     const thorClientEmptyTest = new Client({
-//       phrase,
-//       network: 'testnet' as Network /*, derivationPath: "44'/931'/0'/0/0"*/,
-//     })
-//     const addressTest = thorClientEmptyTest.getAddress()
-//     expect(addressTest).toEqual(testnet_address_path0)
+    const thorClientEmptyTest = new Client({
+      phrase,
+      network: 'testnet' as Network /*, derivationPath: "44'/931'/0'/0/0"*/,
+    })
+    const addressTest = thorClientEmptyTest.getAddress()
+    expect(addressTest).toEqual(testnet_address_path0)
 
-//     const viaSetPhraseAddr1Test = thorClientEmptyTest.getAddress(1 /*, "44'/931'/0'/0/1"*/)
-//     expect(viaSetPhraseAddr1Test).toEqual(testnet_address_path1)
+    const viaSetPhraseAddr1Test = thorClientEmptyTest.getAddress(1 /*, "44'/931'/0'/0/1"*/)
+    expect(viaSetPhraseAddr1Test).toEqual(testnet_address_path1)
 
-//     const thorClientEmptyMain1 = new Client({
-//       phrase,
-//       network: 'mainnet' as Network /*, derivationPath: "44'/931'/0'/0/1"*/,
-//     })
-//     const addressMain1 = thorClientEmptyMain1.getAddress(1)
-//     expect(addressMain1).toEqual(mainnet_address_path1)
+    const thorClientEmptyMain1 = new Client({
+      phrase,
+      network: 'mainnet' as Network /*, derivationPath: "44'/931'/0'/0/1"*/,
+    })
+    const addressMain1 = thorClientEmptyMain1.getAddress(1)
+    expect(addressMain1).toEqual(mainnet_address_path1)
 
-//     const thorClientEmptyTest1 = new Client({
-//       phrase,
-//       network: 'testnet' as Network /*, derivationPath: "44'/931'/0'/0/1"*/,
-//     })
-//     const addressTest1 = thorClientEmptyTest1.getAddress(1)
-//     expect(addressTest1).toEqual(testnet_address_path1)
-//   })
+    const thorClientEmptyTest1 = new Client({
+      phrase,
+      network: 'testnet' as Network /*, derivationPath: "44'/931'/0'/0/1"*/,
+    })
+    const addressTest1 = thorClientEmptyTest1.getAddress(1)
+    expect(addressTest1).toEqual(testnet_address_path1)
+  })
 
-//   it('throws an error passing an invalid phrase', async () => {
-//     expect(() => {
-//       new Client({ phrase: 'invalid phrase', network: 'mainnet' as Network })
-//     }).toThrow()
+  it('throws an error passing an invalid phrase', async () => {
+    expect(() => {
+      new Client({ phrase: 'invalid phrase', network: 'mainnet' as Network })
+    }).toThrow()
 
-//     expect(() => {
-//       new Client({ phrase: 'invalid phrase', network: 'testnet' as Network })
-//     }).toThrow()
-//   })
+    expect(() => {
+      new Client({ phrase: 'invalid phrase', network: 'testnet' as Network })
+    }).toThrow()
+  })
 
-//   it('should have right address', async () => {
-//     expect(thorClient.getAddress()).toEqual(testnet_address_path0)
+  it('should have right address', async () => {
+    expect(thorClient.getAddress()).toEqual(testnet_address_path0)
 
-//     expect(thorMainClient.getAddress()).toEqual(mainnet_address_path0)
-//   })
+    expect(thorMainClient.getAddress()).toEqual(mainnet_address_path0)
+  })
 
-//   it('should allow to get the CosmosSDKClient', async () => {
-//     expect(thorClient.getCosmosClient()).toBeInstanceOf(CosmosSDKClient)
-//   })
+  it('should allow to get the CosmosSDKClient', async () => {
+    expect(thorClient.getCosmosClient()).toBeInstanceOf(CosmosSDKClient)
+  })
 
-//   it('should update net', async () => {
-//     thorMainClient.setNetwork('testnet' as Network)
-//     expect(thorMainClient.getNetwork()).toEqual('testnet')
+  it('should update net', async () => {
+    thorMainClient.setNetwork('testnet' as Network)
+    expect(thorMainClient.getNetwork()).toEqual('testnet')
 
-//     const address = await thorMainClient.getAddress()
-//     expect(address).toEqual(testnet_address_path0)
-//   })
+    const address = await thorMainClient.getAddress()
+    expect(address).toEqual(testnet_address_path0)
+  })
 
-//   it('should init, should have right prefix', async () => {
-//     expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
+  it('should init, should have right prefix', async () => {
+    expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
 
-//     thorClient.setNetwork('mainnet' as Network)
-//     expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
-//   })
+    thorClient.setNetwork('mainnet' as Network)
+    expect(thorClient.validateAddress(thorClient.getAddress())).toEqual(true)
+  })
 
-//   it('should have right client url', async () => {
-//     thorClient.setClientUrl({
-//       mainnet: {
-//         node: 'new mainnet client',
-//         rpc: 'new mainnet client',
-//       },
-//       testnet: {
-//         node: 'new testnet client',
-//         rpc: 'new testnet client',
-//       },
-//     })
+  it('should have right client url', async () => {
+    thorClient.setClientUrl({
+      mainnet: {
+        node: 'new mainnet client',
+        rpc: 'new mainnet client',
+      },
+      testnet: {
+        node: 'new testnet client',
+        rpc: 'new testnet client',
+      },
+    })
 
-//     thorClient.setNetwork('mainnet' as Network)
-//     expect(thorClient.getClientUrl().node).toEqual('new mainnet client')
+    thorClient.setNetwork('mainnet' as Network)
+    expect(thorClient.getClientUrl().node).toEqual('new mainnet client')
 
-//     thorClient.setNetwork('testnet' as Network)
-//     expect(thorClient.getClientUrl().node).toEqual('new testnet client')
-//   })
+    thorClient.setNetwork('testnet' as Network)
+    expect(thorClient.getClientUrl().node).toEqual('new testnet client')
+  })
 
-//   it('returns private key', async () => {
-//     const privKey = thorClient.getPrivKey()
-//     expect(privKey.toBase64()).toEqual('CHCbyYWorMZVRFtfJzt72DigvZeRNi3jo2c3hGEQ46I=')
-//   })
+  it('returns private key', async () => {
+    const privKey = thorClient.getPrivKey()
+    expect(Buffer.from(privKey.bytes()).toString('base64')).toEqual('CHCbyYWorMZVRFtfJzt72DigvZeRNi3jo2c3hGEQ46I=')
+  })
 
-//   it('returns public key', async () => {
-//     const pubKey = thorClient.getPubKey()
-//     expect(pubKey.toBase64()).toEqual('AsL4F+rvFMqDkZYpVVnZa0OBa0EXwscjNrODbBME42vC')
-//   })
+  it('returns public key', async () => {
+    const pubKey = thorClient.getPubKey()
+    const pkString = Buffer.from(pubKey.bytes()).toString('base64')
+    expect(pkString).toEqual('AsL4F+rvFMqDkZYpVVnZa0OBa0EXwscjNrODbBME42vC')
+  })
 
-//   it('has no balances', async () => {
-//     mockAccountsBalance(thorClient.getClientUrl().node, testnet_address_path0, {
-//       height: 0,
-//       result: [],
-//     })
-//     const result = await thorClient.getBalance(thorClient.getAddress(0))
-//     expect(result).toEqual([])
-//   })
+  it('has no balances', async () => {
+    mockAccountsBalance(thorClient.getClientUrl().node, testnet_address_path0, {
+      balances: [],
+    })
+    const result = await thorClient.getBalance(thorClient.getAddress(0))
+    expect(result).toEqual([])
+  })
 
-//   it('has balances', async () => {
-//     thorMainClient.setNetwork('mainnet' as Network)
-//     // mainnet - has balance: thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5
-//     // mainnet - 0: thor19kacmmyuf2ysyvq3t9nrl9495l5cvktjs0yfws
-//     mockAccountsBalance(thorMainClient.getClientUrl().node, 'thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5', {
-//       height: 0,
-//       result: [
-//         {
-//           denom: 'rune',
-//           amount: '100',
-//         },
-//       ],
-//     })
+  it('has balances', async () => {
+    thorMainClient.setNetwork('mainnet' as Network)
+    // mainnet - has balance: thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5
+    // mainnet - 0: thor19kacmmyuf2ysyvq3t9nrl9495l5cvktjs0yfws
+    mockAccountsBalance(thorMainClient.getClientUrl().node, 'thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5', {
+      balances: [
+        {
+          denom: 'rune',
+          amount: '100',
+        },
+      ],
+    })
 
-//     const balances = await thorMainClient.getBalance('thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5')
-//     expect(balances.length).toEqual(1)
-//     expect(balances[0].asset).toEqual(AssetRuneNative)
-//     expect(balances[0].amount.amount().isEqualTo(baseAmount(100).amount())).toBeTruthy()
-//   })
+    const balances = await thorMainClient.getBalance('thor147jegk6e9sum7w3svy3hy4qme4h6dqdkgxhda5')
+    expect(balances.length).toEqual(1)
+    expect(balances[0].asset).toEqual(AssetRuneNative)
+    expect(balances[0].amount.amount().isEqualTo(baseAmount(100).amount())).toBeTruthy()
+  })
 
-//   it('has an empty tx history', async () => {
-//     const expected: TxsPage = {
-//       total: 0,
-//       txs: [],
-//     }
+  // it('rune + synth balances', async () => {
+  //   thorClient.setNetwork(Network.Testnet)
+  //   mockAccountsBalance(thorClient.getClientUrl().node, 'tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg', {
+  //     balances: [
+  //       {
+  //         denom: 'bnb/bnb',
+  //         amount: '100',
+  //       },
+  //       {
+  //         denom: 'bnb/busd-74e',
+  //         amount: '200',
+  //       },
+  //       {
+  //         denom: 'rune',
+  //         amount: '200',
+  //       },
+  //     ],
+  //   })
 
-//     mockTxHistory(thorClient.getClientUrl().rpc, {
-//       jsonrpc: '2.0',
-//       id: -1,
-//       result: {
-//         txs: [],
-//         total_count: '0',
-//       },
-//     })
+  //   const balances = await thorClient.getBalance('tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg')
+  //   expect(balances.length).toEqual(3)
+  //   // BNB synth
+  //   expect(balances[0].asset).toEqual({ ...AssetBNB, synth: true })
+  //   expect(balances[0].amount.amount().isEqualTo(baseAmount(100).amount()))
+  //   // BUSD synth
+  //   expect(balances[1].asset).toEqual({ chain: 'BNB', symbol: 'BUSD-74E', ticker: 'BUSD', synth: true })
+  //   expect(balances[1].amount.amount().isEqualTo(baseAmount(200).amount()))
+  //   // RUNE
+  //   expect(balances[2].asset).toEqual(AssetRuneNative)
+  //   expect(balances[2].amount.amount().isEqualTo(baseAmount(300).amount()))
+  // })
 
-//     const transactions = await thorClient.getTransactions({
-//       address: 'tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg',
-//       limit: 1,
-//     })
-//     expect(transactions).toEqual(expected)
-//   })
+  // it('filter BUSD synth balances', async () => {
+  //   const BUSD_ASSET_SYNTH: Asset = { chain: BNBChain, symbol: 'BUSD-74E', ticker: 'BUSD', synth: true }
+  //   thorClient.setNetwork(Network.Testnet)
+  //   mockAccountsBalance(thorClient.getClientUrl().node, 'tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg', {
+  //     balances: [
+  //       {
+  //         denom: 'bnb/bnb',
+  //         amount: '100',
+  //       },
+  //       {
+  //         denom: 'bnb/busd-74e',
+  //         amount: '200',
+  //       },
+  //       {
+  //         denom: 'rune',
+  //         amount: '200',
+  //       },
+  //     ],
+  //   })
 
-//   it('has tx history', async () => {
-//     const historyData = require('../__mocks__/responses/tx_search/sender-tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f.json')
-//     const bondTxData = require('../__mocks__/responses/txs/bond-tn-9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C.json')
-//     const address = 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f'
-//     const txHash = '9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C'
-//     mockTxHistory(thorClient.getClientUrl().rpc, historyData)
+  //   const balances = await thorClient.getBalance('tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg')
+  //   expect(balances.length).toEqual(1)
+  //   // BUSD synth
+  //   expect(balances[0].asset).toEqual(BUSD_ASSET_SYNTH)
+  //   expect(balances[0].amount.amount().isEqualTo(baseAmount(200).amount()))
+  // })
 
-//     assertTxHashGet(thorClient.getClientUrl().node, txHash, bondTxData)
+  it('has an empty tx history', async () => {
+    const expected: TxsPage = {
+      total: 0,
+      txs: [],
+    }
 
-//     const txs = await thorClient.getTransactions({
-//       address: 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f',
-//     })
+    mockTxHistory(thorClient.getClientUrl().rpc, {
+      jsonrpc: '2.0',
+      id: -1,
+      result: {
+        txs: [],
+        total_count: '0',
+      },
+    })
 
-//     expect(txs.total).toEqual(1)
+    const transactions = await thorClient.getTransactions({
+      address: 'tthor13gym97tmw3axj3hpewdggy2cr288d3qffr8skg',
+      limit: 1,
+    })
+    expect(transactions).toEqual(expected)
+  })
 
-//     const { type, hash, asset, from, to } = txs.txs[0]
+  it('has tx history', async () => {
+    const historyData = require('../__mocks__/responses/tx_search/sender-tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f.json')
+    const bondTxData = require('../__mocks__/responses/txs/bond-tn-9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C.json')
+    const address = 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f'
+    const txHash = '9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C'
+    mockTxHistory(thorClient.getClientUrl().rpc, historyData)
 
-//     expect(type).toEqual('transfer')
-//     expect(hash).toEqual(txHash)
-//     expect(asset).toEqual(AssetRuneNative)
-//     expect(from[0].from).toEqual(address)
-//     expect(from[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
-//     expect(from[1].from).toEqual(address)
-//     expect(from[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
-//     expect(to[0].to).toEqual('tthor1dheycdevq39qlkxs2a6wuuzyn4aqxhve3hhmlw')
-//     expect(to[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
-//     expect(to[1].to).toEqual('tthor17gw75axcnr8747pkanye45pnrwk7p9c3uhzgff')
-//     expect(to[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
-//   })
+    assertTxHashGet(thorClient.getClientUrl().node, txHash, bondTxData)
 
-//   it('transfer', async () => {
-//     const to_address = 'tthor19kacmmyuf2ysyvq3t9nrl9495l5cvktj5c4eh4'
-//     const send_amount: BaseAmount = baseAmount(10000, 6)
-//     const memo = 'transfer'
+    const txs = await thorClient.getTransactions({
+      address: 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f',
+    })
 
-//     const expected_txsPost_result = {
-//       check_tx: {},
-//       deliver_tx: {},
-//       txhash: 'EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D',
-//       height: 0,
-//       logs: [],
-//     }
+    expect(txs.total).toEqual(1)
 
-//     mockAccountsAddress(thorClient.getClientUrl().node, testnet_address_path0, {
-//       height: 0,
-//       result: {
-//         coins: [
-//           {
-//             denom: 'rune',
-//             amount: '210000000',
-//           },
-//         ],
-//         account_number: '0',
-//         sequence: '0',
-//       },
-//     })
-//     mockAccountsBalance(thorClient.getClientUrl().node, testnet_address_path0, {
-//       height: 0,
-//       result: [
-//         {
-//           denom: 'rune',
-//           amount: '210000000',
-//         },
-//       ],
-//     })
-//     assertTxsPost(thorClient.getClientUrl().node, memo, expected_txsPost_result)
+    const { type, hash, asset, from, to } = txs.txs[0]
 
-//     const result = await thorClient.transfer({
-//       asset: AssetRuneNative,
-//       recipient: to_address,
-//       amount: send_amount,
-//       memo,
-//     })
+    expect(type).toEqual('transfer')
+    expect(hash).toEqual(txHash)
+    expect(asset).toEqual(AssetRuneNative)
+    expect(from[0].from).toEqual(address)
+    expect(from[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
+    expect(from[1].from).toEqual(address)
+    expect(from[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
+    expect(to[0].to).toEqual('tthor1dheycdevq39qlkxs2a6wuuzyn4aqxhve3hhmlw')
+    expect(to[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
+    expect(to[1].to).toEqual('tthor17gw75axcnr8747pkanye45pnrwk7p9c3uhzgff')
+    expect(to[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
+  })
 
-//     expect(result).toEqual('EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D')
-//   })
+  it('transfer', async () => {
+    const to_address = 'tthor19kacmmyuf2ysyvq3t9nrl9495l5cvktj5c4eh4'
+    const send_amount: BaseAmount = baseAmount(10000, 6)
+    const memo = 'transfer'
 
-//   it('deposit', async () => {
-//     const send_amount: BaseAmount = baseAmount(10000, 6)
-//     const memo = 'swap:BNB.BNB:tbnb1ftzhmpzr4t8ta3etu4x7nwujf9jqckp3th2lh0'
+    const expected_txsPost_result = {
+      tx_response: {
+        txhash: 'EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D',
+        code: 0,
+      },
+    }
 
-//     const expected_txsPost_result = {
-//       check_tx: {},
-//       deliver_tx: {},
-//       txhash: 'EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D',
-//       height: 0,
-//       logs: [],
-//     }
+    mockAccountsAddress(thorClient.getClientUrl().node, testnet_address_path0, {
+      account: {
+        '@type': '/cosmos.auth.v1beta1.BaseAccount',
+        address: testnet_address_path0,
+        pub_key: {
+          '@type': '/cosmos.crypto.secp256k1.PubKey',
+          key: 'AyB84hKBjN2wsmdC2eF1Ppz6l3VxlfSKJpYsTaL4VrrE',
+        },
+        account_number: '0',
+        sequence: '0',
+      },
+    })
+    mockAccountsBalance(thorClient.getClientUrl().node, testnet_address_path0, {
+      balances: [
+        {
+          denom: 'rune',
+          amount: '210000000',
+        },
+      ],
+    })
+    assertTxsPost(thorClient.getClientUrl().node, expected_txsPost_result)
 
-//     mockAccountsAddress(thorClient.getClientUrl().node, testnet_address_path0, {
-//       height: 0,
-//       result: {
-//         coins: [
-//           {
-//             denom: 'rune',
-//             amount: '210000000',
-//           },
-//         ],
-//         account_number: '0',
-//         sequence: '0',
-//       },
-//     })
-//     mockAccountsBalance(thorClient.getClientUrl().node, testnet_address_path0, {
-//       height: 0,
-//       result: [
-//         {
-//           denom: 'rune',
-//           amount: '210000000',
-//         },
-//       ],
-//     })
-//     mockThorchainDeposit(thorClient.getClientUrl().node, {
-//       type: 'cosmos-sdk/StdTx',
-//       value: {
-//         msg: [
-//           {
-//             type: 'thorchain/MsgDeposit',
-//             value: {
-//               coins: [
-//                 {
-//                   asset: 'THOR.RUNE',
-//                   amount: '10000',
-//                 },
-//               ],
-//               memo: 'swap:BNB.BNB:tbnb1ftzhmpzr4t8ta3etu4x7nwujf9jqckp3th2lh0',
-//               signer: 'tthor19kacmmyuf2ysyvq3t9nrl9495l5cvktj5c4eh4',
-//             },
-//           },
-//         ],
-//         fee: {
-//           amount: [],
-//           gas: '100000000',
-//         },
-//         signatures: [],
-//         memo: '',
-//         timeout_height: '0',
-//       },
-//     })
-//     assertTxsPost(thorClient.getClientUrl().node, '', expected_txsPost_result)
+    const result = await thorClient.transfer({
+      asset: AssetRuneNative,
+      recipient: to_address,
+      amount: send_amount,
+      memo,
+    })
 
-//     const result = await thorClient.deposit({
-//       asset: AssetRuneNative,
-//       amount: send_amount,
-//       memo,
-//     })
+    expect(result).toEqual('EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D')
+  })
 
-//     expect(result).toEqual('EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D')
-//   })
+  it('deposit', async () => {
+    const send_amount: BaseAmount = baseAmount(10000, 6)
+    const memo = 'swap:BNB.BNB:tbnb1ftzhmpzr4t8ta3etu4x7nwujf9jqckp3th2lh0'
 
-//   it('get transaction data for BOND tx', async () => {
-//     const txData = require('../__mocks__/responses/txs/bond-tn-9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C.json')
-//     const txHash = '9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C'
-//     const address = 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f'
-//     assertTxHashGet(thorClient.getClientUrl().node, txHash, txData)
-//     const { type, hash, asset, from, to } = await thorClient.getTransactionData(txHash, address)
+    const expected_txsPost_result = {
+      tx_response: {
+        txhash: 'EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D',
+        code: 0,
+      },
+    }
 
-//     expect(type).toEqual('transfer')
-//     expect(hash).toEqual(txHash)
-//     expect(asset).toEqual(AssetRuneNative)
-//     expect(from[0].from).toEqual(address)
-//     expect(from[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
-//     expect(from[1].from).toEqual(address)
-//     expect(from[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
-//     expect(to[0].to).toEqual('tthor1dheycdevq39qlkxs2a6wuuzyn4aqxhve3hhmlw')
-//     expect(to[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
-//     expect(to[1].to).toEqual('tthor17gw75axcnr8747pkanye45pnrwk7p9c3uhzgff')
-//     expect(to[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
-//   })
+    const nodeUrl = thorClient.getClientUrl().node
 
-//   it('should return valid explorer url', () => {
-//     expect(thorClient.getExplorerUrl()).toEqual('https://viewblock.io/thorchain?network=testnet')
+    mockAccountsAddress(nodeUrl, testnet_address_path0, {
+      account: {
+        '@type': '/cosmos.auth.v1beta1.BaseAccount',
+        address: testnet_address_path0,
+        pub_key: {
+          '@type': '/cosmos.crypto.secp256k1.PubKey',
+          key: 'AyB84hKBjN2wsmdC2eF1Ppz6l3VxlfSKJpYsTaL4VrrE',
+        },
+        account_number: '0',
+        sequence: '0',
+      },
+    })
+    mockAccountsBalance(nodeUrl, testnet_address_path0, {
+      balances: [
+        {
+          denom: 'rune',
+          amount: '210000000',
+        },
+      ],
+    })
+    mockTendermintNodeInfo(nodeUrl, {
+      default_node_info: {
+        network: 'chain-id-testnet',
+      },
+    })
+    mockThorchainConstants(nodeUrl)
 
-//     thorClient.setNetwork('mainnet' as Network)
-//     expect(thorClient.getExplorerUrl()).toEqual('https://viewblock.io/thorchain')
-//   })
+    assertTxsPost(nodeUrl, expected_txsPost_result)
 
-//   it('should return valid explorer address url', () => {
-//     expect(thorClient.getExplorerAddressUrl('tthorabc')).toEqual(
-//       'https://viewblock.io/thorchain/address/tthorabc?network=testnet',
-//     )
+    const result = await thorClient.deposit({
+      asset: AssetRuneNative,
+      amount: send_amount,
+      memo,
+    })
 
-//     thorClient.setNetwork('mainnet' as Network)
-//     expect(thorClient.getExplorerAddressUrl('thorabc')).toEqual('https://viewblock.io/thorchain/address/thorabc')
-//   })
+    expect(result).toEqual('EA2FAC9E82290DCB9B1374B4C95D7C4DD8B9614A96FACD38031865EB1DBAE24D')
+  })
 
-//   it('should return valid explorer tx url', () => {
-//     expect(thorClient.getExplorerTxUrl('txhash')).toEqual('https://viewblock.io/thorchain/tx/txhash?network=testnet')
+  it('get transaction data for BOND tx', async () => {
+    const txData = require('../__mocks__/responses/txs/bond-tn-9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C.json')
+    const txHash = '9C175AF7ACE9FCDC930B78909FFF598C18CBEAF9F39D7AA2C4D9A27BB7E55A5C'
+    const address = 'tthor137kees65jmhjm3gxyune0km5ea0zkpnj4lw29f'
+    assertTxHashGet(thorClient.getClientUrl().node, txHash, txData)
+    const { type, hash, asset, from, to } = await thorClient.getTransactionData(txHash, address)
 
-//     thorClient.setNetwork('mainnet' as Network)
-//     expect(thorClient.getExplorerTxUrl('txhash')).toEqual('https://viewblock.io/thorchain/tx/txhash')
-//   })
-// })
+    expect(type).toEqual('transfer')
+    expect(hash).toEqual(txHash)
+    expect(asset).toEqual(AssetRuneNative)
+    expect(from[0].from).toEqual(address)
+    expect(from[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
+    expect(from[1].from).toEqual(address)
+    expect(from[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
+    expect(to[0].to).toEqual('tthor1dheycdevq39qlkxs2a6wuuzyn4aqxhve3hhmlw')
+    expect(to[0].amount.amount().toString()).toEqual(assetToBase(assetAmount(0.02)).amount().toString())
+    expect(to[1].to).toEqual('tthor17gw75axcnr8747pkanye45pnrwk7p9c3uhzgff')
+    expect(to[1].amount.amount().toString()).toEqual(assetToBase(assetAmount(1700)).amount().toString())
+  })
+
+  it('should return valid explorer url', () => {
+    expect(thorClient.getExplorerUrl()).toEqual('https://viewblock.io/thorchain?network=testnet')
+
+    thorClient.setNetwork(Network.Mainnet)
+    expect(thorClient.getExplorerUrl()).toEqual('https://viewblock.io/thorchain')
+  })
+
+  it('should return valid explorer address url', () => {
+    expect(thorClient.getExplorerAddressUrl('tthorabc')).toEqual(
+      'https://viewblock.io/thorchain/address/tthorabc?network=testnet',
+    )
+
+    thorClient.setNetwork(Network.Mainnet)
+    expect(thorClient.getExplorerAddressUrl('thorabc')).toEqual('https://viewblock.io/thorchain/address/thorabc')
+  })
+
+  it('should return valid explorer tx url', () => {
+    expect(thorClient.getExplorerTxUrl('txhash')).toEqual('https://viewblock.io/thorchain/tx/txhash?network=testnet')
+
+    thorClient.setNetwork('mainnet' as Network)
+    expect(thorClient.getExplorerTxUrl('txhash')).toEqual('https://viewblock.io/thorchain/tx/txhash')
+  })
+  it('fetches fees from client', async () => {
+    const url = thorClient.getClientUrl().node
+    mockThorchainConstants(url)
+
+    const fees = await thorClient.getFees()
+
+    expect(fees.average.amount().toString()).toEqual('2000000')
+    expect(fees.fast.amount().toString()).toEqual('2000000')
+    expect(fees.fastest.amount().toString()).toEqual('2000000')
+  })
+
+  it('returns default fees if client is not available', async () => {
+    const url = thorClient.getClientUrl().node
+    nock(url).get('/thorchain/constants').reply(404)
+
+    const fees = await thorClient.getFees()
+
+    expect(fees.average.amount().toString()).toEqual('2000000')
+    expect(fees.fast.amount().toString()).toEqual('2000000')
+    expect(fees.fastest.amount().toString()).toEqual('2000000')
+  })
+})